change file structures a bit
