+++ src/components/pod-graph/cytoscape/graph-logic/computation.js
@@ -1,33 +1,5 @@
 import { v4 as uuid } from 'uuid';
-
-export const findSharedCategoriesAndKeywords = (podcast1, podcast2) => removeDuplicateElements([
-  ...podcast1.categories.filter(category => podcast2.categories.includes(category)),
-  ...podcast1.categories.filter(category => podcast2.keywords.includes(category)),
-  ...podcast1.keywords.filter(keyword => podcast2.keywords.includes(keyword)),
-  ...podcast1.keywords.filter(keyword => podcast2.categories.includes(keyword)),
-]);
-
-export const haveSharedElements = (arr1, arr2) => arr1.some(item => arr2.includes(item));
-
-/* @return [Array] The given `array`, omitting duplicate as well as falsy elements */
-const removeDuplicateElements = array => [...new Set(array.filter(x => x))];
-
-/* @return [<<Object>>] The findAllDisjointGraphs result mapped onto the subscriptions metadata */
-const finalizeDisjointGraphsObject = (subscriptions, disjointGraphs) => disjointGraphs
-  .map(graph => graph.map(node => subscriptions.find(subscription => subscription.subscribeUrl ===
-      node.subscribeUrl)));
-
-/* @param [<Object>] subscriptions
-   @return [<<Object>>] An array of graphs grouped by shared keywords & categories, where each
-     graph comprises an array of subscription metadata (nodes) */
-export const groupSubscriptionsBySharedKeywords = subscriptions => {
-  const nodes = subscriptions.map(subscription => ({
-    subscribeUrl: subscription.subscribeUrl,
-    keywordsAndCategories: removeDuplicateElements([
-      ...subscription.keywords, ...subscription.categories]),
-  }));
-  return finalizeDisjointGraphsObject(subscriptions, findAllDisjointGraphs(nodes));
-};
+import { findSharedCategoriesAndKeywords, haveSharedElements, removeDuplicateElements } from './utils';
 
 /* @param [<Object>] nodes
    @param [<<Object>>] disjointGraphs The intermediate result through recursion
@@ -44,7 +16,7 @@ export const findAllDisjointGraphs = (nodes, disjointGraphs = []) => {
   do {
     // eslint-disable-next-line no-loop-func
     relatedPodcast = nodes.find(item => item.visited !== true
-      && haveSharedElements(keywordsAndCategoriesInCommon, item.keywordsAndCategories));
+        && haveSharedElements(keywordsAndCategoriesInCommon, item.keywordsAndCategories));
     if (!relatedPodcast) break;
 
     relatedPodcast.visited = true;
@@ -56,6 +28,23 @@ export const findAllDisjointGraphs = (nodes, disjointGraphs = []) => {
   disjointGraphs.push(graph);
 
   return findAllDisjointGraphs(nodes, disjointGraphs);
+};
+
+/* @return [<<Object>>] The findAllDisjointGraphs result mapped onto the subscriptions metadata */
+const finalizeDisjointGraphsObject = (subscriptions, disjointGraphs) => disjointGraphs
+  .map(graph => graph.map(node => subscriptions.find(subscription => subscription.subscribeUrl ===
+      node.subscribeUrl)));
+
+/* @param [<Object>] subscriptions
+   @return [<<Object>>] An array of graphs grouped by shared keywords & categories, where each
+     graph comprises an array of subscription metadata (nodes) */
+export const groupSubscriptionsBySharedKeywords = subscriptions => {
+  const nodes = subscriptions.map(subscription => ({
+    subscribeUrl: subscription.subscribeUrl,
+    keywordsAndCategories: removeDuplicateElements([
+      ...subscription.keywords, ...subscription.categories]),
+  }));
+  return finalizeDisjointGraphsObject(subscriptions, findAllDisjointGraphs(nodes));
 };
 
 export const generateNodes = disjointGraphs => {
@@ -95,10 +84,10 @@ export const generateNodes = disjointGraphs => {
 export const generateEdges = disjointGraphs => {
   const eachDisjointGraphEdges = disjointGraphs.map(graph => graph
     .reduce((acc, podcast, _, arrayReference) => {
-    // A match is any other podcast that has one same category or keyword
+      // A match is any other podcast that has one same category or keyword
       let matches = arrayReference.filter(({ categories, keywords }) => (
         haveSharedElements(podcast.categories, categories)
-        || haveSharedElements(podcast.keywords, keywords)
+          || haveSharedElements(podcast.keywords, keywords)
       ));
 
       // remove loops (edge from a node to itself)
@@ -116,7 +105,7 @@ export const generateEdges = disjointGraphs => {
 
       return [...acc, ...result];
     }, [])
-    // remove duplicate edges since the graph is undirected.
+  // remove duplicate edges since the graph is undirected.
     .reduce((acc, edge) => (
       acc.some(item => item.data.target === edge.data.source && item.data.source === edge.data.target)
         ? acc
